+++ src/components/Create/components/RewardsList/RewardItem.tsx
@@ -5,7 +5,7 @@ import FormattedAddress from 'components/FormattedAddress'
 import { JuiceVideoThumbnail } from 'components/NftRewards/NftVideo/JuiceVideoThumbnail'
 import TooltipLabel from 'components/TooltipLabel'
 import { useContentType } from 'hooks/ContentType'
-import { NftRewardTier } from 'models/nftRewardTier'
+import { NftRewardTier } from 'models/nftRewards'
 import { ReactNode } from 'react'
 import { fileTypeIsVideo } from 'utils/nftRewards'
 import { prettyUrl } from 'utils/url'
