+++ src/components/Create/components/pages/NftRewards/NftRewardsPage.tsx
@@ -18,7 +18,7 @@ import { CREATE_FLOW } from 'constants/fathomEvents'
 import { FEATURE_FLAGS } from 'constants/featureFlags'
 import { useModal } from 'hooks/Modal'
 import { trackFathomGoal } from 'lib/fathom'
-import { JB721GovernanceType } from 'models/nftRewardTier'
+import { JB721GovernanceType } from 'models/nftRewards'
 import { useContext } from 'react'
 import { useAppSelector } from 'redux/hooks/AppSelector'
 import { useSetCreateFurthestPageReached } from 'redux/hooks/EditingCreateFurthestPageReached'
