+++ src/components/Create/components/pages/ReviewDeploy/components/RewardsReview/RewardsReview.tsx
@@ -2,7 +2,7 @@ import { t } from '@lingui/macro'
 import { Row } from 'antd'
 import { RewardsList } from 'components/Create/components/RewardsList'
 import { FEATURE_FLAGS } from 'constants/featureFlags'
-import { JB721GovernanceType, NftRewardTier } from 'models/nftRewardTier'
+import { JB721GovernanceType, NftRewardTier } from 'models/nftRewards'
 import { useCallback, useMemo } from 'react'
 import { useAppDispatch } from 'redux/hooks/AppDispatch'
 import { useAppSelector } from 'redux/hooks/AppSelector'
