+++ src/components/v2v3/V2V3Project/ManageNftsSection/RedeemNftsModal/RedeemNftCard.tsx
@@ -4,7 +4,7 @@ import {
   IPFSNftRewardTier,
   NftRewardTier,
   NFT_METADATA_CONTRIBUTION_FLOOR_ATTRIBUTES_INDEX,
-} from 'models/nftRewardTier'
+} from 'models/nftRewards'
 import { JB721DelegateToken } from 'models/subgraph-entities/v2/jb-721-delegate-tokens'
 import { useQuery, UseQueryResult } from 'react-query'
 import { cidFromIpfsUri, ipfsGatewayUrl } from 'utils/ipfs'
