+++ src/components/v2v3/V2V3Project/V2V3ProjectSettings/pages/EditNftsPage/EditCollectionDetailsSection.tsx
@@ -7,7 +7,7 @@ import { NftRewardsContext } from 'contexts/NftRewards/NftRewardsContext'
 import { V2V3ProjectContext } from 'contexts/v2v3/Project/V2V3ProjectContext'
 import { useNftCollectionMetadata } from 'hooks/JB721Delegate/NftCollectionMetadata'
 import { useReconfigureNftCollectionMetadata } from 'hooks/v2v3/transactor/ReconfigureNftCollectionMetadata'
-import { NftCollectionMetadata } from 'models/nftRewardTier'
+import { NftCollectionMetadata } from 'models/nftRewards'
 import { useCallback, useContext, useEffect, useMemo, useState } from 'react'
 import { MarketplaceFormFields } from '../../../../shared/FundingCycleConfigurationDrawers/NftDrawer/shared/formFields'
 import { NftCollectionDetailsFormItems } from '../../../../shared/FundingCycleConfigurationDrawers/NftDrawer/shared/NftCollectionDetailsFormItems'
