+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer/shared/NftCollectionDetailsFormItems.tsx
@@ -6,7 +6,7 @@ import ExternalLink from 'components/ExternalLink'
 import { MinimalCollapse } from 'components/MinimalCollapse'
 import { RadioItem } from 'components/RadioItem'
 import { FEATURE_FLAGS } from 'constants/featureFlags'
-import { JB721GovernanceType } from 'models/nftRewardTier'
+import { JB721GovernanceType } from 'models/nftRewards'
 import { useMemo } from 'react'
 import { featureFlagEnabled } from 'utils/featureFlags'
 import { helpPagePath } from 'utils/routes'
