+++ src/components/v2v3/shared/FundingCycleConfigurationDrawers/NftDrawer/shared/formFields.ts
@@ -1,4 +1,4 @@
-import { JB721GovernanceType } from 'models/nftRewardTier'
+import { JB721GovernanceType } from 'models/nftRewards'
 
 export type MarketplaceFormFields = {
   collectionName: string
