+++ src/constants/delegateVersions.ts
@@ -1,4 +1,4 @@
-import { JB721DELEGATE_V1, JB721DELEGATE_V1_1 } from 'models/nftRewardTier'
+import { JB721DELEGATE_V1, JB721DELEGATE_V1_1 } from 'models/nftRewards'
 
 export const JB721_DELEGATE_V1: JB721DELEGATE_V1 = '1'
 export const JB721_DELEGATE_V1_1: JB721DELEGATE_V1_1 = '1.1'
