+++ src/contexts/NftRewards/JB721DelegateContracts/JB721DelegateContractsContext.ts
@@ -1,4 +1,5 @@
 import { Contract } from '@ethersproject/contracts'
+import { JB721DelegateVersion } from 'models/nftRewards'
 import { createContext } from 'react'
 
 export interface JB721DelegateContracts {
@@ -10,11 +11,12 @@ export interface JB721DelegateContractsLoading {
   JB721TieredDelegateStoreLoading?: boolean
 }
 
-export const JB721DelegateContractsContext: React.Context<{
+export const JB721DelegateContractsContext = createContext<{
   contracts: JB721DelegateContracts
   loading: JB721DelegateContractsLoading
-  version?: string
-}> = createContext({
+  version: JB721DelegateVersion | undefined
+}>({
   contracts: {},
   loading: {},
+  version: undefined,
 })
