+++ src/contexts/NftRewards/NftRewardsContext.ts
@@ -1,4 +1,4 @@
-import { JB721GovernanceType } from 'models/nftRewardTier'
+import { JB721GovernanceType } from 'models/nftRewards'
 import { createContext } from 'react'
 import {
   DEFAULT_NFT_FLAGS,
