+++ src/hooks/JB721Delegate/JB721DelegateVersion/JB721DelegateVersion.ts
@@ -2,7 +2,7 @@ import {
   JB721_DELEGATE_V1,
   JB721_DELEGATE_V1_1,
 } from 'constants/delegateVersions'
-import { DelegateVersion } from 'models/nftRewardTier'
+import { JB721DelegateVersion } from 'models/nftRewards'
 import { useIsJB721DelegateV1 } from './IsJB721DelegateV1'
 import { useIsJB721DelegateV1_1 } from './IsJB721DelegateV1_1'
 
@@ -10,7 +10,7 @@ export function useJB721DelegateVersion({
   dataSourceAddress,
 }: {
   dataSourceAddress: string | undefined
-}): DelegateVersion | undefined {
+}): JB721DelegateVersion | undefined {
   const { value: isV1 } = useIsJB721DelegateV1({ dataSourceAddress })
   const { value: isV1_1 } = useIsJB721DelegateV1_1({ dataSourceAddress })
