+++ src/hooks/v2v3/transactor/ReconfigureNftCollectionMetadata.ts
@@ -3,7 +3,7 @@ import { ProjectMetadataContext } from 'contexts/shared/ProjectMetadataContext'
 import { TransactionContext } from 'contexts/Transaction/TransactionContext'
 import { useJB721TieredDelegate } from 'hooks/JB721Delegate/contracts/JB721TieredDelegate'
 import { TransactorInstance } from 'hooks/Transactor'
-import { NftCollectionMetadata } from 'models/nftRewardTier'
+import { NftCollectionMetadata } from 'models/nftRewards'
 import { useContext } from 'react'
 import { ipfsUri } from 'utils/ipfs'
 import { pinNftCollectionMetadata } from 'utils/nftRewards'
