+++ src/models/subgraph-entities/v2/jb-721-delegate-tokens.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from '@ethersproject/bignumber'
-import { JB721GovernanceType } from 'models/nftRewardTier'
+import { JB721GovernanceType } from 'models/nftRewards'
 import {
   parseBigNumberKeyVals,
   parseSubgraphEntitiesFromJson,
